+++ packages/mason_logger/test/src/progress_test.dart
@@ -11,12 +11,12 @@ void main() {
   group('Progress', () {
     late Stdout stdout;
     late Stdin stdin;
-    late Stdout stderr;
+    late Level level;
 
     setUp(() {
       stdout = MockStdout();
       stdin = MockStdin();
-      stderr = MockStdout();
+      level = Level.info;
     });
 
     group('.complete', () {
@@ -25,7 +25,7 @@ void main() {
           () async {
             const time = '(0.1s)';
             const message = 'test message';
-            final progress = Progress(message, stdout, stderr);
+            final progress = Progress(message, stdout, level);
             await Future<void>.delayed(const Duration(milliseconds: 100));
             progress.complete();
             verify(
@@ -45,7 +45,34 @@ void main() {
           },
           stdout: () => stdout,
           stdin: () => stdin,
-          stderr: () => stderr,
+        );
+      });
+
+      test('does not write lines to stdout when Level > info', () async {
+        await StdioOverrides.runZoned(
+          () async {
+            const time = '(0.1s)';
+            const message = 'test message';
+            final progress = Progress(message, stdout, Level.warning);
+            await Future<void>.delayed(const Duration(milliseconds: 100));
+            progress.complete();
+            verifyNever(
+              () {
+                stdout.write(
+                  '''${lightGreen.wrap('\b${'\b' * (message.length + 4 + time.length)}⠙')} $message... ${darkGray.wrap(time)}''',
+                );
+              },
+            );
+            verifyNever(
+              () {
+                stdout.write(
+                  '''\b${'\b' * (message.length + 4 + time.length)}\u001b[2K${lightGreen.wrap('✓')} $message ${darkGray.wrap(time)}\n''',
+                );
+              },
+            );
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
         );
       });
     });
@@ -56,7 +83,7 @@ void main() {
           () async {
             const update = 'update';
             const time = '(0.1s)';
-            final progress = Progress('message', stdout, stderr);
+            final progress = Progress('message', stdout, level);
             await Future<void>.delayed(const Duration(milliseconds: 100));
             progress.update(update);
             await Future<void>.delayed(const Duration(milliseconds: 100));
@@ -70,18 +97,39 @@ void main() {
           },
           stdout: () => stdout,
           stdin: () => stdin,
-          stderr: () => stderr,
+        );
+      });
+
+      test('does not writes to stdout when Level > info', () async {
+        await StdioOverrides.runZoned(
+          () async {
+            const update = 'update';
+            const time = '(0.1s)';
+            final progress = Progress('message', stdout, Level.warning);
+            await Future<void>.delayed(const Duration(milliseconds: 100));
+            progress.update(update);
+            await Future<void>.delayed(const Duration(milliseconds: 100));
+            verifyNever(
+              () {
+                stdout.write(
+                  '''${lightGreen.wrap('\b${'\b' * (update.length + time.length + 4)}⠹')} $update... ${darkGray.wrap(time)}''',
+                );
+              },
+            );
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
         );
       });
     });
 
     group('.fail', () {
-      test('writes lines to stderr', () async {
+      test('writes lines to stdout', () async {
         await StdioOverrides.runZoned(
           () async {
             const time = '(0.1s)';
             const message = 'test message';
-            final progress = Progress(message, stdout, stderr);
+            final progress = Progress(message, stdout, level);
             await Future<void>.delayed(const Duration(milliseconds: 100));
             progress.fail();
             verify(
@@ -93,7 +141,7 @@ void main() {
             ).called(1);
             verify(
               () {
-                stderr.write(
+                stdout.write(
                   '''\b${'\b' * (message.length + 4 + time.length)}\u001b[2K${red.wrap('✗')} $message ${darkGray.wrap(time)}\n''',
                 );
               },
@@ -101,7 +149,34 @@ void main() {
           },
           stdout: () => stdout,
           stdin: () => stdin,
-          stderr: () => stderr,
+        );
+      });
+
+      test('does not write to stdout when Level > info', () async {
+        await StdioOverrides.runZoned(
+          () async {
+            const time = '(0.1s)';
+            const message = 'test message';
+            final progress = Progress(message, stdout, Level.warning);
+            await Future<void>.delayed(const Duration(milliseconds: 100));
+            progress.fail();
+            verifyNever(
+              () {
+                stdout.write(
+                  '''${lightGreen.wrap('\b${'\b' * (message.length + 4 + time.length)}⠙')} $message... ${darkGray.wrap(time)}''',
+                );
+              },
+            );
+            verifyNever(
+              () {
+                stdout.write(
+                  '''\b${'\b' * (message.length + 4 + time.length)}\u001b[2K${red.wrap('✗')} $message ${darkGray.wrap(time)}\n''',
+                );
+              },
+            );
+          },
+          stdout: () => stdout,
+          stdin: () => stdin,
         );
       });
     });
@@ -112,7 +187,7 @@ void main() {
           () async {
             const time = '(0.1s)';
             const message = 'test message';
-            final progress = Progress(message, stdout, stderr);
+            final progress = Progress(message, stdout, level);
             await Future<void>.delayed(const Duration(milliseconds: 100));
             progress.cancel();
             verify(
@@ -130,39 +205,32 @@ void main() {
           },
           stdout: () => stdout,
           stdin: () => stdin,
-          stderr: () => stderr,
         );
       });
-    });
 
-    group('.call', () {
-      test('writes lines to stdout', () async {
+      test('does not write to stdout when Level > info', () async {
         await StdioOverrides.runZoned(
           () async {
             const time = '(0.1s)';
             const message = 'test message';
-            final progress = Progress(message, stdout, stderr);
+            final progress = Progress(message, stdout, Level.warning);
             await Future<void>.delayed(const Duration(milliseconds: 100));
-            // ignore: deprecated_member_use_from_same_package
-            progress();
-            verify(
+            progress.cancel();
+            verifyNever(
               () {
                 stdout.write(
                   '''${lightGreen.wrap('\b${'\b' * (message.length + 4 + time.length)}⠙')} $message... ${darkGray.wrap(time)}''',
                 );
               },
-            ).called(1);
-            verify(
-              () {
-                stdout.write(
-                  '''\b${'\b' * (message.length + 4 + time.length)}\u001b[2K${lightGreen.wrap('✓')} $message ${darkGray.wrap(time)}\n''',
-                );
-              },
-            ).called(1);
+            );
+            verifyNever(
+              () => stdout.write(
+                '\b${'\b' * (message.length + 4 + time.length)}\u001b[2K',
+              ),
+            );
           },
           stdout: () => stdout,
           stdin: () => stdin,
-          stderr: () => stderr,
         );
       });
     });